minecraft_agent/requirements.txt
# WebSocket 服务器
websockets>=12.0

# OpenAI SDK（兼容 DeepSeek / 硅基流动 API）
openai>=1.30.0

# 向量数据库
chromadb>=0.5.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据处理
numpy>=1.26.0
pydantic>=2.5.0

# 日志
loguru>=0.7.0

# 异步支持
aiofiles>=23.2.0
